update delete client orm

+++ src/sat_orm/pipeline_orm/client.py
@@ -25,6 +25,8 @@ from sqlalchemy.orm import relationship, validates
 
 # Local Application Import
 from sat_orm.pipeline_orm.pipeline_base import Base
+from sat_orm.pipeline_orm.warehouse import Warehouse
+from sat_orm.pipeline_orm.industrial_athlete import IndustrialAthlete
 
 class Client(Base):
     __tablename__ = 'client'
@@ -119,6 +121,8 @@ def insert_or_update(session, data):
     if client_in_table:
         data.pop('id', None)
         data.pop('client_id', None)
+        data['enable_processing'] = data['enableProcessing']
+        data.pop('enableProcessing', None)
         session.query(Client).filter_by(id=client_id).update(data)
         session.commit()
         return client_id
@@ -139,19 +143,69 @@ def delete(session, data):
             session: sqlalchemy.orm.session.Session
             data: {key: value} dictionary
     '''
+    response = {}
     client_id = data['client_id']
+
+    client_has_warehouse = has_warehouse(session, client_id)
+    if client_has_warehouse:
+        response['error'] = "Client has warehouse"
+        response['message'] = "Ensure that client has no athlete/warehouse before deleting"
+        return response
+    
+    client_has_athlete = has_athlete(session, client_id)
+    if client_has_athlete:
+        response['error'] = "Client has athlete"
+        response['message'] = "Ensure that client has no athlete/warehouse before deleting"
+        return response
+    
     no_of_deleted_rows = session.query(Client).filter_by(id=client_id).delete()
     # session.delete(client)
     if no_of_deleted_rows == 1:
         session.commit()
+    else:
+        response['error'] = "Error deleting client"
+    
+    return response
+    
+
+def has_warehouse(session, client_id):
+    '''
+        Description
+            Checks if there is any warehouse associated with the client.
+
+        params
+            session: sqlalchemy.orm.session.Session
+            client_id: Integer
+
+        returns
+            True: When there is a warehouse which belongs to the client
+            False: When there is no warehouse which belongs to the client
+    '''
+    client_has_warehouse = session.query(Warehouse).filter_by(client_id=client_id).first()
+    if client_has_warehouse:
         return True
     else:
         return False
-    
-    # return {}
 
-    
 
+def has_athlete(session, client_id):
+    '''
+        Description
+            Checks if there is any athlete associated with the client.
+
+        params
+            session: sqlalchemy.orm.session.Session
+            client_id: Integer
+
+        returns
+            True: When there is a athlete which belongs to the client
+            False: When there is no athlete which belongs to the client
+    '''
+    client_has_athlete = session.query(IndustrialAthlete).filter_by(client_id=client_id).first()
+    if client_has_athlete:
+        return True
+    else:
+        return False
